Revise Read me. General cleanup - remove do-nothing overrides. When the order of initialization of the fields of an object is important, then do the initialization in an init().
