--- TaskApplication.xcodeproj/project.xcworkspace/xcshareddata/TaskApplication.xccheckout
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>IDESourceControlProjectFavoriteDictionaryKey</key>
-	<false/>
-	<key>IDESourceControlProjectIdentifier</key>
-	<string>10EB1394-C6B0-462E-85C0-367571C298C7</string>
-	<key>IDESourceControlProjectName</key>
-	<string>TaskApplication</string>
-	<key>IDESourceControlProjectOriginsDictionary</key>
-	<dict>
-		<key>58DEA5F9F78AED03F30276194337F64857474C9E</key>
-		<string>https://github.com/mbcrump/TasksForSwiftWithPersistingData</string>
-		<key>9D491CFCEBD738578C711AFF792F576E6DC8ACC0</key>
-		<string>https://github.com/mbcrump/TasksForSwift.git</string>
-	</dict>
-	<key>IDESourceControlProjectPath</key>
-	<string>TaskApplication.xcodeproj</string>
-	<key>IDESourceControlProjectRelativeInstallPathDictionary</key>
-	<dict>
-		<key>58DEA5F9F78AED03F30276194337F64857474C9E</key>
-		<string>../..</string>
-		<key>9D491CFCEBD738578C711AFF792F576E6DC8ACC0</key>
-		<string>../../../TaskApplication</string>
-	</dict>
-	<key>IDESourceControlProjectURL</key>
-	<string>https://github.com/mbcrump/TasksForSwiftWithPersistingData</string>
-	<key>IDESourceControlProjectVersion</key>
-	<integer>111</integer>
-	<key>IDESourceControlProjectWCCIdentifier</key>
-	<string>58DEA5F9F78AED03F30276194337F64857474C9E</string>
-	<key>IDESourceControlProjectWCConfigurations</key>
-	<array>
-		<dict>
-			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
-			<string>public.vcs.git</string>
-			<key>IDESourceControlWCCIdentifierKey</key>
-			<string>9D491CFCEBD738578C711AFF792F576E6DC8ACC0</string>
-			<key>IDESourceControlWCCName</key>
-			<string>TaskApplication</string>
-		</dict>
-		<dict>
-			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
-			<string>public.vcs.git</string>
-			<key>IDESourceControlWCCIdentifierKey</key>
-			<string>58DEA5F9F78AED03F30276194337F64857474C9E</string>
-			<key>IDESourceControlWCCName</key>
-			<string>TaskAppThatPersistData</string>
-		</dict>
-	</array>
-</dict>
-</plist>
