--- TaskApplication.xcodeproj/project.xcworkspace/xcshareddata/TaskApplication.xcscmblueprint
@@ -1,30 +1,0 @@
-{
-  "DVTSourceControlWorkspaceBlueprintPrimaryRemoteRepositoryKey" : "58DEA5F9F78AED03F30276194337F64857474C9E",
-  "DVTSourceControlWorkspaceBlueprintWorkingCopyRepositoryLocationsKey" : {
-
-  },
-  "DVTSourceControlWorkspaceBlueprintWorkingCopyStatesKey" : {
-    "9D491CFCEBD738578C711AFF792F576E6DC8ACC0" : 0,
-    "58DEA5F9F78AED03F30276194337F64857474C9E" : 0
-  },
-  "DVTSourceControlWorkspaceBlueprintIdentifierKey" : "10EB1394-C6B0-462E-85C0-367571C298C7",
-  "DVTSourceControlWorkspaceBlueprintWorkingCopyPathsKey" : {
-    "9D491CFCEBD738578C711AFF792F576E6DC8ACC0" : "TaskApplication",
-    "58DEA5F9F78AED03F30276194337F64857474C9E" : "TaskAppThatPersistData"
-  },
-  "DVTSourceControlWorkspaceBlueprintNameKey" : "TaskApplication",
-  "DVTSourceControlWorkspaceBlueprintVersion" : 204,
-  "DVTSourceControlWorkspaceBlueprintRelativePathToProjectKey" : "TaskApplication.xcodeproj",
-  "DVTSourceControlWorkspaceBlueprintRemoteRepositoriesKey" : [
-    {
-      "DVTSourceControlWorkspaceBlueprintRemoteRepositoryURLKey" : "https:\/\/github.com\/mbcrump\/TasksForSwiftWithPersistingData",
-      "DVTSourceControlWorkspaceBlueprintRemoteRepositorySystemKey" : "com.apple.dt.Xcode.sourcecontrol.Git",
-      "DVTSourceControlWorkspaceBlueprintRemoteRepositoryIdentifierKey" : "58DEA5F9F78AED03F30276194337F64857474C9E"
-    },
-    {
-      "DVTSourceControlWorkspaceBlueprintRemoteRepositoryURLKey" : "https:\/\/github.com\/mbcrump\/TasksForSwift.git",
-      "DVTSourceControlWorkspaceBlueprintRemoteRepositorySystemKey" : "com.apple.dt.Xcode.sourcecontrol.Git",
-      "DVTSourceControlWorkspaceBlueprintRemoteRepositoryIdentifierKey" : "9D491CFCEBD738578C711AFF792F576E6DC8ACC0"
-    }
-  ]
-}
