--- TaskApplication.xcodeproj/xcuserdata/crump.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>SchemeUserState</key>
-	<dict>
-		<key>TaskApplication.xcscheme</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>0</integer>
-		</dict>
-	</dict>
-	<key>SuppressBuildableAutocreation</key>
-	<dict>
-		<key>86FF5BAC1A6D65FB00A017A1</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-		<key>86FF5BC31A6D65FB00A017A1</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-	</dict>
-</dict>
-</plist>
