+++ TaskApplication/Base.lproj/Main.storyboard
@@ -12,7 +12,7 @@
         <!--Tasks-->
         <scene sceneID="hNz-n2-bh7">
             <objects>
-                <viewController id="9pv-A4-QxB" customClass="FirstViewController" customModule="TaskApplication" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="9pv-A4-QxB" customClass="TableViewController" customModule="TaskApplication" customModuleProvider="target" sceneMemberID="viewController">
                     <layoutGuides>
                         <viewControllerLayoutGuide type="top" id="oVW-MD-uw6"/>
                         <viewControllerLayoutGuide type="bottom" id="wkA-S9-WdJ"/>
@@ -60,7 +60,7 @@
         <!--Add Tasks-->
         <scene sceneID="wg7-f3-ORb">
             <objects>
-                <viewController id="8rJ-Kc-sve" customClass="SecondViewController" customModule="TaskApplication" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="8rJ-Kc-sve" customClass="TaskDetailViewController" customModule="TaskApplication" customModuleProvider="target" sceneMemberID="viewController">
                     <layoutGuides>
                         <viewControllerLayoutGuide type="top" id="Ivv-jf-WiS"/>
                         <viewControllerLayoutGuide type="bottom" id="QkW-A0-wuB"/>
